agent/requirements.txt
fastapi
uvicorn
google-generativeai
mcp
httpx
pydantic
